+++ gates/repo.ts
@@ -39,7 +39,7 @@ allow("LICENSE", 4000);
 allow("flake.nix", 1500);
 allow("bend2/base.bend", 32000);
 allow("bend2/bend.lean", 400000);
-allow("bend2/bend.ts", 42000);
+allow("bend2/bend.ts", 44000);
 allow("bend2/comp.ts", 64000);
 allow("bend2/main.ts", 10000);
 allow(/^bend2\/effs\/[a-z0-9_]+\.(c|js)$/, 4000);